requirements.txt
Flask==3.0.2
twilio==6.50.1
python-dotenv
huggingface_hub
streamlit 
pillow
PyPDF2
